requirements.txt
torch
torchvision
pandas
numpy
pillow
gradio
plotly